+++ src/dicebear/__init__.py
@@ -25,4 +25,4 @@ from .errors import *
 from .models import *
 from .utility import *
 
-__version__ = "2.7.4"
+__version__ = "2.7.5"