+++ src/dicebear/metadata.py
@@ -1,0 +1,281 @@
+Metadata = {
+    "adventurer": {
+        "title": "Adventurer",
+        "creator": "Lisa Wischofsky",
+        "source": "https://www.figma.com/community/file/1184595184137881796",
+        "homepage": "https://www.instagram.com/lischi_art/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "adventurer-neutral": {
+        "title": "Adventurer Neutral",
+        "creator": "Lisa Wischofsky",
+        "source": "https://www.figma.com/community/file/1184595184137881796",
+        "homepage": "https://www.instagram.com/lischi_art/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "avataaars": {
+        "title": "Avataaars",
+        "creator": "Pablo Stanley",
+        "source": "https://avataaars.com/",
+        "homepage": "https://twitter.com/pablostanley",
+        "license": {
+            "name": "Free for personal and commercial use",
+            "url": "https://avataaars.com/",
+        },
+    },
+    "avataaars-neutral": {
+        "title": "Avataaars",
+        "creator": "Pablo Stanley",
+        "source": "https://avataaars.com/",
+        "homepage": "https://twitter.com/pablostanley",
+        "license": {
+            "name": "Free for personal and commercial use.",
+            "url": "https://avataaars.com/",
+        },
+    },
+    "big-ears": {
+        "title": "Face Generator",
+        "creator": "The Visual Team",
+        "source": "https://www.figma.com/community/file/986078800058673824",
+        "homepage": "https://thevisual.team/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "big-ears-neutral": {
+        "title": "Face Generator",
+        "creator": "The Visual Team",
+        "source": "https://www.figma.com/community/file/986078800058673824",
+        "homepage": "https://thevisual.team/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "big-smile": {
+        "title": "Custom Avatar",
+        "creator": "Ashley Seo",
+        "source": "https://www.figma.com/community/file/881358461963645496",
+        "homepage": "http://www.ashleyseo.com/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "bottts": {
+        "title": "Bottts",
+        "creator": "Pablo Stanley",
+        "source": "https://bottts.com/",
+        "homepage": "https://twitter.com/pablostanley",
+        "license": {
+            "name": "Free for personal and commercial use",
+            "url": "https://bottts.com/",
+        },
+    },
+    "bottts-neutral": {
+        "title": "Bottts",
+        "creator": "Pablo Stanley",
+        "source": "https://bottts.com/",
+        "homepage": "https://twitter.com/pablostanley",
+        "license": {
+            "name": "Free for personal and commercial use",
+            "url": "https://bottts.com/",
+        },
+    },
+    "croodles": {
+        "title": "Croodles - Doodle your face",
+        "creator": "vijay verma",
+        "source": "https://www.figma.com/community/file/966199982810283152",
+        "homepage": "https://vjy.me/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "croodles-neutral": {
+        "title": "Croodles - Doodle your face",
+        "creator": "vijay verma",
+        "source": "https://www.figma.com/community/file/966199982810283152",
+        "homepage": "https://vjy.me/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "fun-emoji": {
+        "title": "Fun Emoji Set",
+        "creator": "Davis Uche",
+        "source": "https://www.figma.com/community/file/968125295144990435",
+        "homepage": "https://www.instagram.com/davedirect3/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "icons": {
+        "title": "Bootstrap Icons",
+        "creator": "The Bootstrap Authors",
+        "source": "https://github.com/twbs/icons",
+        "homepage": "https://getbootstrap.com/",
+        "license": {
+            "name": "MIT",
+            "url": "https://github.com/twbs/icons/blob/main/LICENSE",
+        },
+    },
+    "identicon": {
+        "title": "Identicon",
+        "creator": "DiceBear",
+        "source": "https://www.dicebear.com",
+        "homepage": "https://www.dicebear.com",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "initials": {
+        "title": "Initials",
+        "creator": "DiceBear",
+        "source": "https://github.com/dicebear/dicebear",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "lorelei": {
+        "title": "Lorelei",
+        "creator": "Lisa Wischofsky",
+        "source": "https://www.figma.com/community/file/1198749693280469639",
+        "homepage": "https://www.instagram.com/lischi_art/",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "lorelei-neutral": {
+        "title": "Lorelei Neutral",
+        "creator": "Lisa Wischofsky",
+        "source": "https://www.figma.com/community/file/1198749693280469639",
+        "homepage": "https://www.instagram.com/lischi_art/",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "micah": {
+        "title": "Avatar Illustration System",
+        "creator": "Micah Lanier",
+        "source": "https://www.figma.com/community/file/829741575478342595",
+        "homepage": "https://dribbble.com/micahlanier",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "miniavs": {
+        "title": "Miniavs - Free Avatar Creator",
+        "creator": "Webpixels",
+        "source": "https://www.figma.com/community/file/923211396597067458",
+        "homepage": "https://webpixels.io/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "notionists": {
+        "title": "Notionists",
+        "creator": "Zoish",
+        "source": "https://heyzoish.gumroad.com/l/notionists",
+        "homepage": "https://bio.link/heyzoish",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "notionists-neutral": {
+        "title": "Notionists",
+        "creator": "Zoish",
+        "source": "https://heyzoish.gumroad.com/l/notionists",
+        "homepage": "https://bio.link/heyzoish",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "open-peeps": {
+        "title": "Open Peeps",
+        "creator": "Pablo Stanley",
+        "source": "https://www.openpeeps.com/",
+        "homepage": "https://twitter.com/pablostanley",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "personas": {
+        "title": "Personas by Draftbit",
+        "creator": "Draftbit - draftbit.com",
+        "source": "https://personas.draftbit.com/",
+        "homepage": "https://draftbit.com/",
+        "license": {
+            "name": "CC BY 4.0",
+            "url": "https://creativecommons.org/licenses/by/4.0/",
+        },
+    },
+    "pixel-art": {
+        "title": "Pixel Art",
+        "creator": "DiceBear",
+        "source": "https://www.figma.com/community/file/1198754108850888330",
+        "homepage": "https://www.dicebear.com",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "pixel-art-neutral": {
+        "title": "Pixel Art Neutral",
+        "creator": "DiceBear",
+        "source": "https://www.figma.com/community/file/1198754108850888330",
+        "homepage": "https://www.dicebear.com",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "rings": {
+        "title": "Rings",
+        "creator": "DiceBear",
+        "source": "https://www.dicebear.com",
+        "homepage": "https://www.dicebear.com",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "shapes": {
+        "title": "Shapes",
+        "creator": "DiceBear",
+        "source": "https://www.dicebear.com",
+        "homepage": "https://www.dicebear.com",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+    "thumbs": {
+        "title": "Thumbs",
+        "creator": "DiceBear",
+        "source": "https://www.dicebear.com",
+        "homepage": "https://www.dicebear.com",
+        "license": {
+            "name": "CC0 1.0",
+            "url": "https://creativecommons.org/publicdomain/zero/1.0/",
+        },
+    },
+}